+++ lib/src/widgets/toolbar.dart
@@ -1,8 +1,11 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
+import 'package:i18n_extension/i18n_widget.dart';
 
 import '../models/documents/attribute.dart';
+import '../models/themes/quill_dialog_theme.dart';
+import '../models/themes/quill_icon_theme.dart';
 import '../utils/media_pick_setting.dart';
 import 'controller.dart';
 import 'toolbar/arrow_indicated_button_list.dart';
@@ -14,6 +17,7 @@ import 'toolbar/image_button.dart';
 import 'toolbar/indent_button.dart';
 import 'toolbar/insert_embed_button.dart';
 import 'toolbar/link_style_button.dart';
+import 'toolbar/select_alignment_button.dart';
 import 'toolbar/select_header_style_button.dart';
 import 'toolbar/toggle_check_list_button.dart';
 import 'toolbar/toggle_style_button.dart';
@@ -29,6 +33,7 @@ export 'toolbar/insert_embed_button.dart';
 export 'toolbar/link_style_button.dart';
 export 'toolbar/quill_dropdown_button.dart';
 export 'toolbar/quill_icon_button.dart';
+export 'toolbar/select_alignment_button.dart';
 export 'toolbar/select_header_style_button.dart';
 export 'toolbar/toggle_check_list_button.dart';
 export 'toolbar/toggle_style_button.dart';
@@ -57,6 +62,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
     this.color,
     this.filePickImpl,
     this.multiRowsDisplay,
+    this.locale,
     Key? key,
   }) : super(key: key);
 
@@ -65,11 +71,14 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
     double toolbarIconSize = kDefaultIconSize,
     bool showBoldButton = true,
     bool showItalicButton = true,
+    bool showSmallButton = false,
     bool showUnderLineButton = true,
     bool showStrikeThrough = true,
+    bool showInlineCode = true,
     bool showColorButton = true,
     bool showBackgroundColorButton = true,
     bool showClearFormat = true,
+    bool showAlignmentButtons = false,
     bool showHeaderStyle = true,
     bool showListNumbers = true,
     bool showListBullets = true,
@@ -90,19 +99,37 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
     FilePickImpl? filePickImpl,
     WebImagePickImpl? webImagePickImpl,
     WebVideoPickImpl? webVideoPickImpl,
+
+    ///The theme to use for the icons in the toolbar, uses type [QuillIconTheme]
+    QuillIconTheme? iconTheme,
+
+    ///The theme to use for the theming of the [LinkDialog()],
+    ///shown when embedding an image, for example
+    QuillDialogTheme? dialogTheme,
+
+    ///The locale to use for the editor toolbar, defaults to system locale
+    ///Currently the supported locales are:
+    /// * Locale('en')
+    /// * Locale('de')
+    /// * Locale('fr')
+    /// * Locale('zh')
+    Locale? locale,
     Key? key,
   }) {
     final isButtonGroupShown = [
       showHistory ||
           showBoldButton ||
           showItalicButton ||
+          showSmallButton ||
           showUnderLineButton ||
           showStrikeThrough ||
+          showInlineCode ||
           showColorButton ||
           showBackgroundColorButton ||
           showClearFormat ||
           onImagePickCallback != null ||
           onVideoPickCallback != null,
+      showAlignmentButtons,
       showHeaderStyle,
       showListNumbers || showListBullets || showListCheck || showCodeBlock,
       showQuote || showIndent,
@@ -113,6 +140,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
       key: key,
       toolBarHeight: toolbarIconSize * 2,
       multiRowsDisplay: multiRowsDisplay,
+      locale: locale,
       children: [
         if (showHistory)
           HistoryButton(
@@ -120,6 +148,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             undo: true,
+            iconTheme: iconTheme,
           ),
         if (showHistory)
           HistoryButton(
@@ -127,6 +156,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             undo: false,
+            iconTheme: iconTheme,
           ),
         if (showBoldButton)
           ToggleStyleButton(
@@ -134,6 +164,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.format_bold,
             iconSize: toolbarIconSize,
             controller: controller,
+            iconTheme: iconTheme,
           ),
         if (showItalicButton)
           ToggleStyleButton(
@@ -141,6 +172,15 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.format_italic,
             iconSize: toolbarIconSize,
             controller: controller,
+            iconTheme: iconTheme,
+          ),
+        if (showSmallButton)
+          ToggleStyleButton(
+            attribute: Attribute.small,
+            icon: Icons.format_size,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            iconTheme: iconTheme,
           ),
         if (showUnderLineButton)
           ToggleStyleButton(
@@ -148,6 +188,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.format_underline,
             iconSize: toolbarIconSize,
             controller: controller,
+            iconTheme: iconTheme,
           ),
         if (showStrikeThrough)
           ToggleStyleButton(
@@ -155,6 +196,15 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.format_strikethrough,
             iconSize: toolbarIconSize,
             controller: controller,
+            iconTheme: iconTheme,
+          ),
+        if (showInlineCode)
+          ToggleStyleButton(
+            attribute: Attribute.inlineCode,
+            icon: Icons.code,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            iconTheme: iconTheme,
           ),
         if (showColorButton)
           ColorButton(
@@ -162,6 +212,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             background: false,
+            iconTheme: iconTheme,
           ),
         if (showBackgroundColorButton)
           ColorButton(
@@ -169,12 +220,14 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             background: true,
+            iconTheme: iconTheme,
           ),
         if (showClearFormat)
           ClearFormatButton(
             icon: Icons.format_clear,
             iconSize: toolbarIconSize,
             controller: controller,
+            iconTheme: iconTheme,
           ),
         if (showImageButton)
           ImageButton(
@@ -185,6 +238,8 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             filePickImpl: filePickImpl,
             webImagePickImpl: webImagePickImpl,
             mediaPickSettingSelector: mediaPickSettingSelector,
+            iconTheme: iconTheme,
+            dialogTheme: dialogTheme,
           ),
         if (showVideoButton)
           VideoButton(
@@ -195,23 +250,43 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             filePickImpl: filePickImpl,
             webVideoPickImpl: webImagePickImpl,
             mediaPickSettingSelector: mediaPickSettingSelector,
+            iconTheme: iconTheme,
+            dialogTheme: dialogTheme,
           ),
         if ((onImagePickCallback != null || onVideoPickCallback != null) &&
             showCameraButton)
           CameraButton(
-              icon: Icons.photo_camera,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              onImagePickCallback: onImagePickCallback,
-              onVideoPickCallback: onVideoPickCallback,
-              filePickImpl: filePickImpl,
-              webImagePickImpl: webImagePickImpl,
-              webVideoPickImpl: webVideoPickImpl),
+            icon: Icons.photo_camera,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            onImagePickCallback: onImagePickCallback,
+            onVideoPickCallback: onVideoPickCallback,
+            filePickImpl: filePickImpl,
+            webImagePickImpl: webImagePickImpl,
+            webVideoPickImpl: webVideoPickImpl,
+            iconTheme: iconTheme,
+          ),
         if (isButtonGroupShown[0] &&
             (isButtonGroupShown[1] ||
                 isButtonGroupShown[2] ||
                 isButtonGroupShown[3] ||
-                isButtonGroupShown[4]))
+                isButtonGroupShown[4] ||
+                isButtonGroupShown[5]))
+          VerticalDivider(
+            indent: 12,
+            endIndent: 12,
+            color: Colors.grey.shade400,
+          ),
+        if (showAlignmentButtons)
+          SelectAlignmentButton(
+            controller: controller,
+            iconSize: toolbarIconSize,
+          ),
+        if (isButtonGroupShown[1] &&
+            (isButtonGroupShown[2] ||
+                isButtonGroupShown[3] ||
+                isButtonGroupShown[4] ||
+                isButtonGroupShown[5]))
           VerticalDivider(
             indent: 12,
             endIndent: 12,
@@ -221,11 +296,12 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
           SelectHeaderStyleButton(
             controller: controller,
             iconSize: toolbarIconSize,
+            iconTheme: iconTheme,
           ),
-        if (isButtonGroupShown[1] &&
-            (isButtonGroupShown[2] ||
-                isButtonGroupShown[3] ||
-                isButtonGroupShown[4]))
+        if (isButtonGroupShown[2] &&
+            (isButtonGroupShown[3] ||
+                isButtonGroupShown[4] ||
+                isButtonGroupShown[5]))
           VerticalDivider(
             indent: 12,
             endIndent: 12,
@@ -237,6 +313,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             icon: Icons.format_list_numbered,
             iconSize: toolbarIconSize,
+            iconTheme: iconTheme,
           ),
         if (showListBullets)
           ToggleStyleButton(
@@ -244,6 +321,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             icon: Icons.format_list_bulleted,
             iconSize: toolbarIconSize,
+            iconTheme: iconTheme,
           ),
         if (showListCheck)
           ToggleCheckListButton(
@@ -251,6 +329,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             icon: Icons.check_box,
             iconSize: toolbarIconSize,
+            iconTheme: iconTheme,
           ),
         if (showCodeBlock)
           ToggleStyleButton(
@@ -258,9 +337,10 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             icon: Icons.code,
             iconSize: toolbarIconSize,
+            iconTheme: iconTheme,
           ),
-        if (isButtonGroupShown[2] &&
-            (isButtonGroupShown[3] || isButtonGroupShown[4]))
+        if (isButtonGroupShown[3] &&
+            (isButtonGroupShown[4] || isButtonGroupShown[5]))
           VerticalDivider(
             indent: 12,
             endIndent: 12,
@@ -272,6 +352,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             controller: controller,
             icon: Icons.format_quote,
             iconSize: toolbarIconSize,
+            iconTheme: iconTheme,
           ),
         if (showIndent)
           IndentButton(
@@ -279,6 +360,7 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             isIncrease: true,
+            iconTheme: iconTheme,
           ),
         if (showIndent)
           IndentButton(
@@ -286,8 +368,9 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             isIncrease: false,
+            iconTheme: iconTheme,
           ),
-        if (isButtonGroupShown[3] && isButtonGroupShown[4])
+        if (isButtonGroupShown[4] && isButtonGroupShown[5])
           VerticalDivider(
             indent: 12,
             endIndent: 12,
@@ -297,12 +380,15 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
           LinkStyleButton(
             controller: controller,
             iconSize: toolbarIconSize,
+            iconTheme: iconTheme,
+            dialogTheme: dialogTheme,
           ),
         if (showHorizontalRule)
           InsertEmbedButton(
             controller: controller,
             icon: Icons.horizontal_rule,
             iconSize: toolbarIconSize,
+            iconTheme: iconTheme,
           ),
       ],
     );
@@ -320,23 +406,34 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
 
   final FilePickImpl? filePickImpl;
 
+  ///The locale to use for the editor toolbar, defaults to system locale
+  ///Currently the supported locales are:
+  /// * Locale('en')
+  /// * Locale('de')
+  /// * Locale('fr')
+  /// * Locale('zh', 'CN')
+  final Locale? locale;
+
   @override
   Size get preferredSize => Size.fromHeight(toolBarHeight);
 
   @override
   Widget build(BuildContext context) {
-    if (multiRowsDisplay ?? true) {
-      return Wrap(
-        alignment: WrapAlignment.center,
-        runSpacing: 4,
-        spacing: 4,
-        children: children,
-      );
-    }
-    return Container(
-      constraints: BoxConstraints.tightFor(height: preferredSize.height),
-      color: color ?? Theme.of(context).canvasColor,
-      child: ArrowIndicatedButtonList(buttons: children),
+    return I18n(
+      initialLocale: locale,
+      child: multiRowsDisplay ?? true
+          ? Wrap(
+              alignment: WrapAlignment.center,
+              runSpacing: 4,
+              spacing: 4,
+              children: children,
+            )
+          : Container(
+              constraints:
+                  BoxConstraints.tightFor(height: preferredSize.height),
+              color: color ?? Theme.of(context).canvasColor,
+              child: ArrowIndicatedButtonList(buttons: children),
+            ),
     );
   }
 }