+++ CHANGELOG.md
@@ -1,3 +1,69 @@
+## [2.0.14]
+* enable customize the checkbox widget using DefaultListBlockStyle style.
+
+## [2.0.13]
+* Improve the scrolling performance by reducing the repaint areas.
+
+## [2.0.12]
+* Fix the selection effect can't be seen as the textLine with background color.
+
+## [2.0.11]
+* Fix visibility of text selection handlers on scroll.
+
+## [2.0.10]
+* cursorConnt.color notify the text_line to repaint if it was disposed.
+
+## [2.0.9]
+* Improve UX when trying to add a link.
+
+## [2.0.8]
+* Adding translations to the toolbar.
+
+## [2.0.7]
+* Added theming options for toolbar icons and LinkDialog.
+
+## [2.0.6]
+* Avoid runtime error when placed inside TabBarView.
+
+## [2.0.5]
+* Support inline code formatting.
+
+## [2.0.4]
+* Enable history shortcuts for desktop.
+
+## [2.0.3]
+* Fix cursor when line contains image.
+
+## [2.0.2]
+* Address KeyboardListener class name conflict.
+
+## [2.0.1]
+* Upgrade flutter_colorpicker to 0.5.0.
+
+## [2.0.0]
+* Text Alignment functions + Block Format standards.
+
+## [1.9.6]
+* Support putting QuillEditor inside a Scrollable view.
+
+## [1.9.5]
+* Skip image when pasting.
+
+## [1.9.4]
+* Bug fix for cursor position when tapping at the end of line with image(s).
+
+## [1.9.3]
+* Bug fix when line only contains one image.
+
+## [1.9.2]
+* Support for building custom inline styles.
+
+## [1.9.1]
+* Cursor jumps to the most appropriate offset to display selection.
+
+## [1.9.0]
+* Support inline image.
+
 ## [1.8.3]
 * Updated quill_delta.